third_party/blink/web_tests/virtual/dialogfocus-old-behavior/external/wpt/html/semantics/interactive-elements/the-dialog-element/dialog-focusing-steps-disconnected-expected.txt

This is a testharness.js-based test.
FAIL dialog.show(): focusing steps should not run on disconnected <dialog> assert_equals: Focusing steps should not run expected Element node <input></input> but got Element node <body>
<input>
<script>
test(() => {
  const outerInput =...
PASS Test that focusing steps do not run when dialog is disconnected
Harness: the test ran to completion.
